Fix tx_starts to match original SpliceAI behavior
The Reference class was not converting the 0-based transcript start from the annotation file to be 1-based. This caused there to be differences in the model outputs due to the distance from transcript start being off by 1 compared to the original code. After changing this to match the original SpliceAI functionality the transcript start positions need adjusted back to 0-based due to the NCLS find_overlap method being start-exclusive.